Initial Public Release
If you know what you are doing, feel free to make edits of this for your own menu. I, and ErectBan, give no permission to use any portion of this code to create a commercial product or any menu that will charge any amount of money to access (This includes paywalls, requiring purchase, requiring donation, requiring Patreon or similar service). This menu is provided for free and all forks of it, no matter how much or how little is used, are to remain that way. Completely free.